requirements.txt
pandas>=2.0.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.12.0
web3>=6.0.0
polygon-api-client>=1.16.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.44.0
requests>=2.32.0







